_base/setup_environment/requirements.txt
bcrypt==4.0.1
boto3==1.26.91
botocore==1.29.91
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
idna==3.4
jmespath==1.0.1
paramiko==3.1.0
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.15
